+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'm.nairn@gmail.com'
 license           'Apache 2.0'
 description      'Installs Vagrant and Vagrant plugins'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.4.0'
+version          '1.5.0'
 
 %w{ mac_os_x ubuntu }.each do |platform|
   supports platform
@@ -17,17 +17,17 @@ end
 recipe 'vagrant::default', "Installs Vagrant and Vagrant plugins"
 
 attribute 'vagrant/download_url',
-          :display_name => "Download URL",
-          :description => "Download URL",
+          :display_name => "Vagrant download URL",
+          :description => "Vagrant download URL",
           :required => "required",
-          :default => "http://files.vagrantup.com/packages",
+          :default => "https://dl.bintray.com/mitchellh/vagrant",
           :recipes => ["vagrant::default"]
 
 attribute 'vagrant/version',
           :display_name => "Vagrant version to install",
           :description => "Vagrant version to install",
           :required => "required",
-          :default => "1.3.5",
+          :default => "1.6.3",
           :recipes => ["vagrant::default"]
 
 attribute 'vagrant/plugins',