+++ CHANGELOG.md
@@ -15,3 +15,9 @@ This file is used to list changes made in each version of vagrant.
 * Updated default Vagrant version to 1.4.3
 * Updated to work with new download endpoint(http://www.vagrantup.com/downloads.html)
 * Add optional version attribute for vagrant plugins
+
+## 1.5.0:
+
+* Updated default Vagrant version to 1.6.3
+* Fix plugin installation
+* Added test kitchen config and basic tests